moved all util functions to utils direcory
