+++ package.json
@@ -21,12 +21,6 @@
     "lodash": "^4.16.4"
   },
   "devDependencies": {
-    "xmldom": "^0.1.22",
-    "nodexml": "^1.0.2",
-    "soap": "^0.13.0",
-    "soap-cookie": "^0.10.1",
-    "vsphere-schema": "^1.0.0",
-    "bluebird": "^3.4.6",
     "chalk": "^1.1.3",
     "chai": "^3.5.0",
     "babel-core": "^6.0.0",
@@ -36,7 +30,6 @@
     "babel-runtime": "^6.9.0",
     "babel-register": "^6.11.0",
     "mocha": "^3.0.0",
-    "rewire": "^2.5.1",
     "rollup": "^0.34.0",
     "rollup-plugin-babel": "^2.6.0",
     "babel-preset-es2015-rollup": "^1.2.0"
