+++ src/client/utils/errorHandler.js
@@ -1,0 +1,4 @@
+export default function errorHandler (err, callback, reject) {
+  callback(err)
+  return reject(err)
+}
