+++ src/client/utils/moRef.js
@@ -1,0 +1,3 @@
+export default function moRef (type, value) {
+  return { type, value }
+}
