+++ src/client/utils/resultHandler.js
@@ -1,0 +1,4 @@
+export default function resultHandler (result, callback, resolve) {
+  callback(null, result)
+  return resolve(result)
+}
